Refactor: 폴더명 변경 및 이동에 따른 import문 업데이트
- GenreNotFoundException를 genre/exception으로 이동
- entity폴더명 entities로 수정 및 엔티티 파일 이동
