+++ src/genre/genre.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Genre } from './entity/genre.entity';
+import { Genre } from './entities/genre.entity';
 import { GenreController } from './genre.controller';
 import { GenreService } from './genre.service';
 import { GenreRepository } from './repository/genre.repository';
